+++ packages/apollo-angular/http/src/types.ts
@@ -22,6 +22,7 @@ export type OperationPrinter = (operation: DocumentNode) => string;
 export interface Options extends FetchOptions, HttpRequestOptions {
   operationPrinter?: OperationPrinter;
   useGETForQueries?: boolean;
+  extractFiles?: ExtractFiles;
 }
 
 export type Body = {
@@ -40,9 +41,7 @@ export type Request = {
   options: HttpRequestOptions;
 };
 
-export type ExtractFiles = (
-  body: Body | Body[],
-) => {
+export type ExtractFiles = (body: Body | Body[]) => {
   clone: Body;
   files: Map<any, any>;
 };