+++ packages/apollo-angular/CHANGELOG.md
@@ -1,14 +1,29 @@
 # Change log
 
-### vNext
+### v3.0.0
 
 - `useGETForQueries` in Http Link [`5996109`](https://github.com/kamilkisiela/apollo-angular/commit/599610934e5f9cf8cc2d62abc270edddc937cc60)
-- Support `TypedDocumentNode` [`120594a`](https://github.com/kamilkisiela/apollo-angular/commit/120594a3f744e0d233ede3239b08e6721cb35d54)
 - Bring back `ApolloModule` [`0a24c4f`](https://github.com/kamilkisiela/apollo-angular/commit/0a24c4f7af8189a825483c57b2d60cb7a875b18f)
 - Fix `useInitialLoading` [`750429c`](https://github.com/kamilkisiela/apollo-angular/commit/750429cbe0aaad07a3fdc2e2ef046fe8be4aad28)
 - Typed `gql` and `graphql` tags - both accept two generic types and support `TypedDocumentNode` [`9a8ea5f`](https://github.com/kamilkisiela/apollo-angular/commit/9a8ea5f229cf7937d74332092cb3eba40828b7b1)
 - Add `useMutationLoading` flag [`bc223fe`](https://github.com/kamilkisiela/apollo-angular/commit/bc223fe6487edd35c56ad908e4739580ce69f056)
 - Fix type inference for Mutations [#1659](https://github.com/kamilkisiela/apollo-angular/pull/1659)
+- Declare support for Angular 13
+- Remove `extract-files` library from dependencies (you need to pass `extractFiles` function to HttpLink's options)
+
+**Migration from v2 to v3**
+
+1. Use `ApolloModule` in your NgModule to provide `Apollo` service
+2. If you're using old version of `graphql`, update to latest v15 or v16.
+3. In case of file uploads, import `extractFiles` from `extract-files` library and pass it to `HttpLink.create({ ..., extractFiles })`
+4. Done.
+
+### v2.6.0
+
+- Support `TypedDocumentNode` [`120594a`](https://github.com/kamilkisiela/apollo-angular/commit/120594a3f744e0d233ede3239b08e6721cb35d54)
+
+### v2.5.0
+
 - Declare support for Angular 12
 
 ### v2.4.0